Abstract logic for currentMonth and prevPeriod into DateUtils
