+++ src/lib/utils/DateUtils.ts
@@ -14,3 +14,36 @@ export const MONTH_NAMES = [
 ];
 
 export const WEEK_DAYS = ['Dom', 'Lun', 'Mar', 'Mié', 'Jue', 'Vie', 'Sáb'];
+
+// utils/dateUtils.ts
+
+export function getCurrentMonthDates(): { firstDay: string; lastDay: string } {
+  const now = new Date();
+  const firstDay = new Date(now.getFullYear(), now.getMonth(), 1)
+    .toISOString()
+    .split('T')[0];
+  const lastDay = new Date(now.getFullYear(), now.getMonth() + 1, 0)
+    .toISOString()
+    .split('T')[0];
+  return { firstDay, lastDay };
+}
+
+export function getPreviousPeriod(
+  start: string,
+  end: string,
+): {
+  startDate: string;
+  endDate: string;
+} {
+  const startDate = new Date(start);
+  const endDate = new Date(end);
+
+  const diffInMs = endDate.getTime() - startDate.getTime();
+  const previousEnd = new Date(startDate.getTime() - 1);
+  const previousStart = new Date(previousEnd.getTime() - diffInMs);
+
+  return {
+    startDate: previousStart.toISOString().split('T')[0],
+    endDate: previousEnd.toISOString().split('T')[0],
+  };
+}
